+++ lib/engineyard/eyrc.rb
@@ -16,6 +16,10 @@ module EY
       path.exist?
     end
 
+    def delete_api_token
+      delete('api_token')
+    end
+
     def api_token
       self['api_token']
     end
@@ -35,17 +39,24 @@ module EY
     end
 
     def []=(key,val)
-      merge_and_write(key.to_s => val)
+      new_data = read_data.merge(key.to_s => val)
+      write_data new_data
       val
+    end
+
+    def delete(key)
+      data = read_data.dup
+      res = data.delete(key)
+      write_data data
+      res
     end
 
     def read_data
       exist? && YAML.load(path.read) || {}
     end
 
-    def merge_and_write(new_data)
-      to_write = read_data.merge(new_data)
-      path.open("w") {|f| YAML.dump(to_write, f) }
+    def write_data(new_data)
+      path.open("w") {|f| YAML.dump(new_data, f) }
     end
 
   end