+++ ChangeLog.md
@@ -2,7 +2,7 @@
 
 ## NEXT
 
-  *
+  * ey logout removes the API token from ~/.eyrc
 
 ## v1.4.20 (2012-01-13)